Update README.md
These movies and series are the last ones that I recently watched.
Movies
Gangnam Blues 
Gangnam Blues is a riveting South Korean crime film set against the backdrop of Seoul's Gangnam district. Its intense storyline and compelling characters make it a must-watch.
The Wolf of Wall Street 
Directed by Martin Scorsese, The Wolf of Wall Street is a wild ride through the world of excess and finance, featuring Leonardo DiCaprio's unforgettable performance as Jordan Belfort.
Margin Call 
Margin Call provides a gripping look into the high-stakes world of investment banking during the 2008 financial crisis. The stellar cast and taut storytelling make it a standout in the genre.
Series 
Tom and Jerry
Tom and Jerry, the classic cat-and-mouse duo, never fails to bring laughter. The timeless antics of these two animated characters have made it a cherished series for generations.
Dora The Explorer 
Join Dora and her monkey friend Boots on exciting adventures! Dora the Explorer combines education with fun, making it an ideal series for children and a nostalgic favorite for many.
Oggy and the Cockroaches
Oggy and the Cockroaches is a hilarious animated series that follows the constant battle between Oggy, a laid-back cat, and three mischievous cockroaches. The slapstick humor and clever storytelling make it a joy to watch.